+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## v0.0.7 (2019-10-07)
+
+ - Add `add file` command with flags `-repo reponame -b branchname -f file -gitname "Author Name" -gitemail "author.email@email.com" -m "Commit message"`
+
 ## v0.0.6 (2019-10-05)
 
  - Add `create branch` command with flags `-repo reponame -b branchname`